Backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-phonenumber-field==7.1.0
django-simple-history==3.3.0
djangorestframework==3.14.0
holidays==0.35
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
phonenumbers==8.13.18
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
